# openapi.yaml
openapi: 3.1.0
info:
  title: Payroo Mini Payrun API
  version: 1.0.0
servers:
  - url: "http://localhost:4000"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    PayrunRequest:
      type: object
      required:
        - periodStart
        - periodEnd
      properties:
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        employeeIds:
          type: array
          items:
            type: string
    Payslip:
      type: object
      properties:
        employeeId:
          type: string
        normalHours:
          type: number
        overtimeHours:
          type: number
        gross:
          type: number
        tax:
          type: number
        super:
          type: number
        net:
          type: number
    Payrun:
      type: object
      properties:
        id:
          type: string
        periodStart:
          type: string
          format: date
        periodEnd:
          type: string
          format: date
        totals:
          type: object
          properties:
            gross:
              type: number
            tax:
              type: number
            super:
              type: number
            net:
              type: number
        payslips:
          type: array
          items:
            $ref: "#/components/schemas/Payslip"
security:
  - bearerAuth: []
paths:
...
  /payruns:
    post:
      summary: Generate a payrun for a period (and optional employee subset)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PayrunRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payrun"
    get:
      summary: List payruns
      responses:
        "200":
          description: OK
  "/payruns/{id}":
    get:
      summary: Get a payrun by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payrun"
...